chore: Move to internal file
